+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Fixes a bug where first anchor was not visualized on creation
   - Adds new class `FlightCollection`, which has an API to set and access active flight Player
 - Possibility to load WMS as single image.
+- Fixes vectorProperties.ts `getValuesForfeatures()` if feature array is empty
 
 ### 5.0.4